Conflicts:
	src/scip/mem.c
	unittests/Makefile

+++ src/scip/cons_and.c
@@ -1042,15 +1042,15 @@ SCIP_RETCODE checkCons(
       {
          solval = SCIPgetSolVal(scip, sol, consdata->vars[i]);
 
-	 /* @todo If "upgraded resultants to varstatus implicit" is fully allowed, than the following assert does not hold
-	  *       anymore, therefor we need to stop the check and return with the status not violated, because the
-	  *       integrality condition of this violated operand needs to be enforced by another constraint.
-	  *
-	  *       The above should be asserted by marking the constraint handler, for which the result needs to be
-	  *       SCIP_SEPARATED if the origin was the CONSENFOPS or the CONSENFOLP callback or SCIP_INFEASIBLE if the
-	  *       origin was CONSCHECK callback.
-	  *
-	  */
+        /* @todo If "upgraded resultants to varstatus implicit" is fully allowed, than the following assert does not hold
+         *       anymore, therefor we need to stop the check and return with the status not violated, because the
+         *       integrality condition of this violated operand needs to be enforced by another constraint.
+         *
+         *       The above should be asserted by marking the constraint handler, for which the result needs to be
+         *       SCIP_SEPARATED if the origin was the CONSENFOPS or the CONSENFOLP callback or SCIP_INFEASIBLE if the
+         *       origin was CONSCHECK callback.
+         *
+         */
          assert(SCIPisFeasIntegral(scip, solval));
          if( solval < 0.5 )
             break;
@@ -3285,7 +3285,7 @@ SCIP_DECL_HASHKEYVAL(hashKeyValAndcons)
    maxidx = SCIPvarGetIndex(consdata->vars[consdata->nvars - 1]);
    assert(minidx >= 0 && minidx <= maxidx);
 
-   hashval = (consdata->nvars << 29) + (minidx << 22) + (mididx << 11) + maxidx; /*lint !e701*/
+   hashval = ((unsigned int)consdata->nvars << 29) + ((unsigned int)minidx << 22) + ((unsigned int)mididx << 11) + maxidx; /*lint !e701*/
 
    return hashval;
 }
@@ -4218,7 +4218,7 @@ SCIP_DECL_CONSENFOLP(consEnfolpAnd)
       {
          if( conshdlrdata->enforcecuts )
          {
-	    SCIP_Bool consseparated;
+            SCIP_Bool consseparated;
 
             SCIP_CALL( separateCons(scip, conss[i], NULL, &consseparated, &cutoff) );
             if ( cutoff )
@@ -4226,18 +4226,18 @@ SCIP_DECL_CONSENFOLP(consEnfolpAnd)
                *result = SCIP_CUTOFF;
                return SCIP_OKAY;
             }
-	    separated = separated || consseparated;
+            separated = separated || consseparated;
 
-	    /* following assert is wrong in the case some variables were not in LP (dynamic columns),
-	     *
-	     * e.g. the resultant, which has a negative objective value, is in the lp solution on its upper bound
-	     * (variables with status loose are in an lp solution on it's best bound), but already creating a row, and
-	     * thereby creating the column, changes the solution value (variable than has status column, and the
-	     * initialization sets the lp solution value) to 0.0, and this already could lead to no violation of the
-	     * rows, which then are not seperated into the lp
-	     */
+            /* following assert is wrong in the case some variables were not in LP (dynamic columns),
+            *
+            * e.g. the resultant, which has a negative objective value, is in the lp solution on its upper bound
+            * (variables with status loose are in an lp solution on it's best bound), but already creating a row, and
+            * thereby creating the column, changes the solution value (variable than has status column, and the
+            * initialization sets the lp solution value) to 0.0, and this already could lead to no violation of the
+            * rows, which then are not seperated into the lp
+            */
 #if 0
-	    assert(consseparated); /* because the solution is integral, the separation always finds a cut */
+            assert(consseparated); /* because the solution is integral, the separation always finds a cut */
 #endif
          }
          else
